move fixtures to separate file
